refac z query fns
