+++ src/main/java/org/delia/db/hls/manager/HLSManager.java
@@ -24,11 +24,11 @@ import org.delia.db.sql.QueryType;
 import org.delia.db.sql.QueryTypeDetector;
 import org.delia.db.sql.fragment.MiniSelectFragmentParser;
 import org.delia.db.sql.fragment.WhereFragmentGenerator;
+import org.delia.queryresponse.LetSpan;
+import org.delia.queryresponse.LetSpanEngine;
 import org.delia.runner.QueryResponse;
 import org.delia.runner.VarEvaluator;
 import org.delia.type.DTypeRegistry;
-import org.delia.zqueryresponse.LetSpan;
-import org.delia.zqueryresponse.LetSpanEngine;
 
 /**
  * Facade between Delia Runner and the db. Allows us to have different strategies
