+++ src/main/java/org/delia/queryresponse/LetSpan.java
@@ -1,4 +1,4 @@
-package org.delia.zqueryresponse;
+package org.delia.queryresponse;
 
 import java.util.ArrayList;
 import java.util.List;
