--- src/main/java/org/delia/queryresponse/QueryResponseFunction.java
@@ -1,8 +1,0 @@
-//package org.delia.queryresponse;
-//
-//import org.delia.compiler.ast.QueryExp;
-//import org.delia.runner.QueryResponse;
-//
-//public interface QueryResponseFunction {
-//	public QueryResponse process(String fnName, QueryExp queryExp, QueryResponse qresp, QueryFuncContext ctx);
-//}
