+++ src/main/java/org/delia/queryresponse/function/ZFetchFunction.java
@@ -1,17 +1,17 @@
-package org.delia.zqueryresponse.function;
+package org.delia.queryresponse.function;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.delia.compiler.ast.QueryFuncExp;
 import org.delia.queryresponse.QueryFuncContext;
+import org.delia.queryresponse.ZQueryResponseFunctionBase;
 import org.delia.runner.FetchRunner;
 import org.delia.runner.QueryResponse;
 import org.delia.type.DRelation;
 import org.delia.type.DTypeRegistry;
 import org.delia.type.DValue;
 import org.delia.util.DValueHelper;
-import org.delia.zqueryresponse.ZQueryResponseFunctionBase;
 
 public class ZFetchFunction extends ZQueryResponseFunctionBase {
 	private FetchRunner fetchRunner;
