+++ src/main/java/org/delia/queryresponse/function/ZMaxFunction.java
@@ -1,24 +1,23 @@
-package org.delia.zqueryresponse.function;
+package org.delia.queryresponse.function;
 
 import java.util.Date;
 import java.util.List;
 
 import org.apache.commons.collections.CollectionUtils;
-import org.delia.compiler.ast.QueryExp;
 import org.delia.compiler.ast.QueryFuncExp;
 import org.delia.core.FactoryService;
 import org.delia.queryresponse.QueryFuncContext;
+import org.delia.queryresponse.ZQueryResponseFunctionBase;
 import org.delia.runner.QueryResponse;
 import org.delia.type.DTypeRegistry;
 import org.delia.type.DValue;
 import org.delia.type.Shape;
 import org.delia.util.DeliaExceptionHelper;
-import org.delia.zqueryresponse.ZQueryResponseFunctionBase;
 
-public class ZMinFunction extends ZQueryResponseFunctionBase {
+public class ZMaxFunction extends ZQueryResponseFunctionBase {
 	private FactoryService factorySvc;
 
-	public ZMinFunction(FactoryService factorySvc, DTypeRegistry registry) {
+	public ZMaxFunction(FactoryService factorySvc, DTypeRegistry registry) {
 		super(registry);
 		this.factorySvc = factorySvc;
 	}
@@ -50,77 +49,77 @@ public class ZMinFunction extends ZQueryResponseFunctionBase {
 		case DATE:
 			return processDate(qresp, dvalList);
 		default:
-			DeliaExceptionHelper.throwError("unsupported-min-type", "min() doesn't support type '%s'", shape);
+			DeliaExceptionHelper.throwError("unsupported-max-type", "max() doesn't support type '%s'", shape);
 		}
 		return qresp;
 	}
 
 	private QueryResponse processInt(QueryResponse qresp, List<DValue> dvalList) {
-		int min = Integer.MAX_VALUE;
+		int max = Integer.MIN_VALUE;
 		for(DValue dval: dvalList) {
 			if (dval == null) {
 				continue;
 			}
 			int k = dval.asInt(); 
-			if (k < min) {
-				min = k;
+			if (k > max) {
+				max = k;
 			}
 		}
 		
-		DValue dval = buildIntVal(min);
+		DValue dval = buildIntVal(max);
 		setSingletonResult(qresp, dval);
 		return qresp;
 	}
 	private QueryResponse processLong(QueryResponse qresp, List<DValue> dvalList) {
-		long min = Long.MAX_VALUE;
+		long max = Integer.MIN_VALUE;
 		for(DValue dval: dvalList) {
 			if (dval == null) {
 				continue;
 			}
-			long k = dval.asLong();
-			if (k < min) {
-				min = k;
+			long k = dval.asLong(); 
+			if (k > max) {
+				max = k;
 			}
 		}
 		
-		DValue dval = buildLongVal(min);
+		DValue dval = buildLongVal(max);
 		setSingletonResult(qresp, dval);
 		return qresp;
 	}
 	private QueryResponse processNumber(QueryResponse qresp, List<DValue> dvalList) {
-		double min = Double.MAX_VALUE;
+		double max = Double.MIN_VALUE;
 		for(DValue dval: dvalList) {
 			if (dval == null) {
 				continue;
 			}
 			double k = dval.asNumber();
-			if (k < min) {
-				min = k;
+			if (k > max) {
+				max = k;
 			}
 		}
 		
-		DValue dval = buildNumberVal(min);
+		DValue dval = buildNumberVal(max);
 		setSingletonResult(qresp, dval);
 		return qresp;
 	}
 	private QueryResponse processBoolean(QueryResponse qresp, List<DValue> dvalList) {
-		Boolean min = true;
+		Boolean max = false;
 		for(DValue dval: dvalList) {
 			if (dval == null) {
 				continue;
 			}
-			Boolean k = dval.asBoolean();
-			if (k.compareTo(min) < 0) {
-				min = k;
+			Boolean k = dval.asBoolean(); 
+			if (k.compareTo(max) > 0) {
+				max = k;
 			}
 		}
 		
-		DValue dval = buildBoolVal(min);
+		DValue dval = buildBoolVal(max);
 		setSingletonResult(qresp, dval);
 		return qresp;
 	}
 	private QueryResponse processString(QueryResponse qresp, List<DValue> dvalList) {
-		String min = null; //max possible string
+		String min = null; //min possible string
 		for(DValue dval: dvalList) {
 			if (dval == null) {
 				continue;
@@ -129,7 +128,7 @@ public class ZMinFunction extends ZQueryResponseFunctionBase {
 			
 			if (min == null) {
 				min = k;
-			} else if (k.compareTo(min) < 0) {
+			} else if (k.compareTo(min) > 0) {
 				min = k;
 			}
 		}
@@ -139,7 +138,7 @@ public class ZMinFunction extends ZQueryResponseFunctionBase {
 		return qresp;
 	}
 	private QueryResponse processDate(QueryResponse qresp, List<DValue> dvalList) {
-		Date min = new Date(Long.MAX_VALUE); //max possible string
+		Date min = new Date(Long.MIN_VALUE);
 		for(DValue dval: dvalList) {
 			if (dval == null) {
 				continue;
@@ -148,7 +147,7 @@ public class ZMinFunction extends ZQueryResponseFunctionBase {
 			
 			if (min == null) {
 				min = k;
-			} else if (k.compareTo(min) < 0) {
+			} else if (k.compareTo(min) > 0) {
 				min = k;
 			}
 		}
