+++ src/test/java/org/delia/mem/FilmAndActor2Tests.java
@@ -5,7 +5,6 @@ import java.util.List;
 
 import org.delia.api.Delia;
 import org.delia.api.DeliaSession;
-import org.delia.base.DBTestHelper;
 import org.delia.bdd.NewBDDBase;
 import org.delia.builder.ConnectionBuilder;
 import org.delia.builder.ConnectionInfo;
